requirements.txt
# Core graph analysis
networkx>=3.1
numpy>=1.24
scipy>=1.10

# Community detection
python-louvain>=0.16
leidenalg>=0.10
igraph>=0.11

# API
fastapi>=0.104
uvicorn[standard]>=0.24
strawberry-graphql[fastapi]>=0.215

# Configuration
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0

# Data processing
pandas>=2.0

# Logging and monitoring
structlog>=23.0

# Optional: Neo4j
# neo4j>=5.0

# Development
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.21
black>=23.0
ruff>=0.1.0
mypy>=1.0
pre-commit>=3.0
